Moved attributes to more sensible namespace
