+++ SoulsFormats/Attributes/RotationRadians.cs
@@ -1,0 +1,9 @@
+﻿using System;
+
+namespace SoulsFormats.Attributes
+{
+    [AttributeUsage(AttributeTargets.Property, AllowMultiple = false)]
+    public class RotationRadians : Attribute
+    {
+    }
+}
