--- SoulsFormats/Util/Attributes.cs
@@ -1,30 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace SoulsFormats
-{
-    [System.AttributeUsage(AttributeTargets.Property, AllowMultiple = false)]
-    public class HideProperty : Attribute
-    {
-    }
-
-    [System.AttributeUsage(AttributeTargets.Property, AllowMultiple = false)]
-    public class RotationRadians : Attribute
-    {
-    }
-
-    [System.AttributeUsage(AttributeTargets.Property, AllowMultiple = false)]
-    public class RotationXZY : Attribute
-    {
-    }
-
-    /// <summary>
-    ///     Properties with this attribute are not used as a reference
-    ///     to obtain render groups from.
-    /// </summary>
-    [System.AttributeUsage(AttributeTargets.Property, AllowMultiple = false)]
-    public class NoRenderGroupInheritance : Attribute
-    {
-    }
-}
